chore: cleanup duplicated functions/package and add godoc
- move GetPlatform() from deploy package to cluster package
- move const ManagedRhods SelfManagedRhods OpenDataHub from deploy to cluster package
- move WaitForDeploymentAvailable() monitoring package to cluster package
- remove monitoring package
- move UpdatePodSecurityRolebinding() from common package to cluster package
- deprecate GetDomain from common package, to only use GetDomain from cluster package.
- remove gvk package, move its GVK to cluster package
- move DeleteExistingSubscription() from deploy package to upgrade package
- do not export getSubscription()
