+++ pkg/common/common.go
@@ -15,10 +15,10 @@ limitations under the License.
 */
 
 // Package common contains utility functions used by different components
+// for cluster related common operations, refer to package cluster
 package common
 
 import (
-	"context"
 	"crypto/sha256"
 	b64 "encoding/base64"
 	"fmt"
@@ -27,47 +27,9 @@ import (
 	"strings"
 
 	"github.com/go-logr/logr"
-	routev1 "github.com/openshift/api/route/v1"
 	"go.uber.org/zap/zapcore"
-	authv1 "k8s.io/api/rbac/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
-
-// UpdatePodSecurityRolebinding update default rolebinding which is created in namespace by manifests
-// being used by different components and sre monitoring.
-func UpdatePodSecurityRolebinding(cli client.Client, serviceAccountsList []string, namespace string) error {
-	foundRoleBinding := &authv1.RoleBinding{}
-	err := cli.Get(context.TODO(), client.ObjectKey{Name: namespace, Namespace: namespace}, foundRoleBinding)
-	if err != nil {
-		return err
-	}
-
-	for _, sa := range serviceAccountsList {
-		// Append serviceAccount if not added already
-		if !subjectExistInRoleBinding(foundRoleBinding.Subjects, sa, namespace) {
-			foundRoleBinding.Subjects = append(foundRoleBinding.Subjects, authv1.Subject{
-				Kind:      authv1.ServiceAccountKind,
-				Name:      sa,
-				Namespace: namespace,
-			})
-		}
-	}
-
-	return cli.Update(context.TODO(), foundRoleBinding)
-}
-
-// Internal function used by UpdatePodSecurityRolebinding()
-// Return whether Rolebinding matching service account and namespace exists or not.
-func subjectExistInRoleBinding(subjectList []authv1.Subject, serviceAccountName, namespace string) bool {
-	for _, subject := range subjectList {
-		if subject.Name == serviceAccountName && subject.Namespace == namespace {
-			return true
-		}
-	}
-
-	return false
-}
 
 // ReplaceStringsInFile replaces variable with value in manifests during runtime.
 func ReplaceStringsInFile(fileName string, replacements map[string]string) error {
@@ -157,17 +119,6 @@ func GetMonitoringData(data string) (string, error) {
 	encodedData := b64.StdEncoding.EncodeToString(hashSum)
 
 	return encodedData, nil
-}
-
-// Use openshift-console namespace to get host domain.
-func GetDomain(cli client.Client, name string, namespace string) (string, error) {
-	consoleRoute := &routev1.Route{}
-	if err := cli.Get(context.TODO(), client.ObjectKey{Name: name, Namespace: namespace}, consoleRoute); err != nil {
-		return "", fmt.Errorf("error getting %s route URL: %w", name, err)
-	}
-	domainIndex := strings.Index(consoleRoute.Spec.Host, ".")
-
-	return consoleRoute.Spec.Host[domainIndex+1:], nil
 }
 
 // to use different mode for logging, e.g. development, production
