+++ pkg/upgrade/uninstallation.go
@@ -6,13 +6,13 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-multierror"
+	ofapiv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	dsci "github.com/opendatahub-io/opendatahub-operator/v2/apis/dscinitialization/v1"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
-	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/deploy"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/metadata/labels"
 )
 
@@ -25,7 +25,7 @@ const (
 // OperatorUninstall deletes all the externally generated resources. This includes monitoring resources and applications
 // installed by KfDef.
 func OperatorUninstall(ctx context.Context, cli client.Client) error {
-	platform, err := deploy.GetPlatform(cli)
+	platform, err := cluster.GetPlatform(cli)
 	if err != nil {
 		return err
 	}
@@ -77,12 +77,12 @@ func OperatorUninstall(ctx context.Context, cli client.Client) error {
 
 	fmt.Printf("Removing operator subscription which in turn will remove installplan\n")
 	subsName := "opendatahub-operator"
-	if platform == deploy.SelfManagedRhods {
+	if platform == cluster.SelfManagedRhods {
 		subsName = "rhods-operator"
-	} else if platform == deploy.ManagedRhods {
+	} else if platform == cluster.ManagedRhods {
 		subsName = "addon-managed-odh"
 	}
-	if err := deploy.DeleteExistingSubscription(cli, operatorNs, subsName); err != nil {
+	if err := DeleteExistingSubscription(cli, operatorNs, subsName); err != nil {
 		return err
 	}
 
@@ -161,4 +161,30 @@ func removeCSV(ctx context.Context, c client.Client) error {
 	}
 	fmt.Printf("No clusterserviceversion for the operator found.\n")
 	return nil
+}
+
+// DeleteExistingSubscription deletes given Subscription if it exists
+// Do not error if the Subscription does not exist.
+func DeleteExistingSubscription(cli client.Client, operatorNs string, subsName string) error {
+	sub, err := getSubscription(cli, operatorNs, subsName)
+	if err != nil {
+		return client.IgnoreNotFound(err)
+	}
+
+	if err := cli.Delete(context.TODO(), sub); client.IgnoreNotFound(err) != nil {
+		return fmt.Errorf("error deleting subscription %s: %w", sub.Name, err)
+	}
+
+	return nil
+}
+
+// GetSubscription checks if a Subscription for the operator exists in the given namespace.
+// if exist, return object; otherwise, return error.
+func getSubscription(cli client.Client, namespace string, name string) (*ofapiv1alpha1.Subscription, error) {
+	sub := &ofapiv1alpha1.Subscription{}
+	if err := cli.Get(context.TODO(), client.ObjectKey{Namespace: namespace, Name: name}, sub); err != nil {
+		// real error or 'not found' both return here
+		return nil, err
+	}
+	return sub, nil
 }
